fix/category

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 
 class CategoryController extends Controller
 {
@@ -30,16 +31,12 @@ class CategoryController extends Controller
     public function store(Request $request)
     {
         $data = $request->validate([
-            'name' => 'required|string|max:255', // Xác thực trường name
-            'description' => 'nullable|string',  // Xác thực trường description
+            'name' => 'required|string|max:255|unique:categories,name', 
         ]);
+        $data['description'] = Str::slug($data['name']);
+        Category::query()->create($data); 
 
-        // Laravel sẽ tự động thêm giá trị cho `created_at` và `updated_at`, nên không cần validate chúng.
-
-        Category::query()->create($data);  // Tạo một bản ghi mới với dữ liệu đã xác thực
-
-        return redirect()->route('admin.categories.index');  // Redirect về trang index của categories (nếu route có đúng tên)
-        // dd($request->all()); // Debug dữ liệu request nếu cần
+        return redirect()->route('admin.categories.index');  
     }
 
 
@@ -65,9 +62,9 @@ class CategoryController extends Controller
     public function update(Request $request, category $category)
     {
         $data = $request->validate([
-            'name' => 'required|string|max:255', // Xác thực trường name
-            'description' => 'nullable|string', // Xác thực trường description
+            'name' => 'required|string|max:255',
         ]);
+        $data['description'] = Str::slug($data['name']);
         $category->update($data);
         return redirect()->route('admin.categories.index');
     }
@@ -77,10 +74,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-
-        // Xóa bản ghi category
+        if($category->productHome()->exists()){
+            return back()->with('error', 'Không thể xóa danh mục đang có liên kết với sản phẩm');
+        }
         $category->delete();
-
         return redirect()->route('admin.categories.index');
     }
 }
